fix: use self'.packages.go for `golangci-lint` (#1)
It makes more sense to use the updated version of go exposed by this
flake rather than nixpkgs one which could potentially be outdated.
Another alternative would be to make `packages` recursive and just use
`go` instead of `self'`, but I tend to prefer going through `self` as it
is more refactor-proof and makes it very explicit where the package
comes from.